moved-components-from-Home-to-a-folder
