+++ myCookingManager/frontend/src/components/HomePage/Home.jsx
